--- src/postgres/db.rs
@@ -1,32 +1,0 @@
-use diesel::pg::PgConnection;
-use diesel::r2d2::ConnectionManager;
-use lazy_static::lazy_static;
-use r2d2;
-use std::env;
-
-type Pool = r2d2::Pool<ConnectionManager<PgConnection>>;
-pub type DbConnection = r2d2::PooledConnection<ConnectionManager<PgConnection>>;
-
-embed_migrations!("../tgbot/src/postgres/migrations");
-
-lazy_static! {
-    static ref POOL: Pool = {
-        let db_url = env::var("DATABASE_URL").expect("Database url not set");
-        let manager = ConnectionManager::<PgConnection>::new(db_url);
-        let builder = r2d2::Pool::builder();
-
-        builder
-            .build(manager)
-            .expect("Failed to create database connection pool")
-    };
-}
-
-pub fn init() {
-    lazy_static::initialize(&POOL);
-    let conn = connection().expect("Failed to get db connection");
-    embedded_migrations::run(&conn).unwrap();
-}
-
-pub fn connection() -> anyhow::Result<DbConnection> {
-    POOL.get().map_err(anyhow::Error::from)
-}
