+++ src/providers/crypto/users/mod.rs
@@ -1,0 +1,2 @@
+pub mod model;
+pub use model::*;
