+++ src/providers/mod.rs
@@ -1,3 +1,2 @@
 pub mod crypto;
 pub mod db;
-pub mod schema;
