+++ src/traits.rs
@@ -1,11 +1,8 @@
-#[async_trait::async_trait]
-pub trait UsersManager {
-    type User;
+pub trait CryptoProvider {
+    type Transaction;
+    type Signature;
     type Error;
-
-    async fn get_all(&self) -> Result<Vec<Self::User>, Self::Error>;
-    async fn find(&self, username: String) -> Result<Self::User, Self::Error>;
-    async fn delete(&self, username: String) -> Result<(), Self::Error>;
-    async fn update(&self, user: Self::User) -> Result<(), Self::Error>;
-    async fn insert(&self, user: Self::User) -> Result<(), Self::Error>;
+    fn generate_keypairs(self) -> Result<String, Self::Error>;
+    fn get_public(self) -> Result<String, Self::Error>;
+    fn sign_transaction(self, transaction: Self::Transaction) -> Result<String, Self::Error>;
 }
